StringIndexOf vs AsciiStringIndexOf Bench Threads1

+++ common/src/main/java/io/netty/util/AsciiStringUtil.java
@@ -32,7 +32,7 @@ public final class AsciiStringUtil {
         final int length = toIndex - fromIndex;
         final int longCount = length >>> 3;
         final long pattern = SWARByteUtil.compilePattern(value);
-        for (int i = 0; i < longCount; ++i) {
+        for (long i = 0; i < longCount; ++i) {
             final long word = PlatformDependent.getLong(bytes, fromIndex);
             final long mask = SWARByteUtil.applyPattern(word, pattern);
             if (mask != 0) {
